combine lockedlist and locked, fix route
