+++ src/containers/QueryApolloProvider.tsx
@@ -5,7 +5,6 @@ import React, { PropsWithChildren } from 'react';
 import { ApolloClient, ApolloLink, ApolloProvider, HttpLink, InMemoryCache } from '@apollo/client';
 import { offsetLimitPagination, getMainDefinition } from '@apollo/client/utilities';
 import { WebSocketLink } from '@apollo/client/link/ws';
-import { PropsWithChildren } from 'react';
 
 const getHttpLink = (uri: string | undefined) => new HttpLink({ uri });
