+++ src/pages/account/Withdrawn/Locked/Home/Locked.module.css
@@ -1,3 +1,22 @@
 .withdrawnContainer {
   margin: 1rem 0;
 }
+
+.container {
+  display: flex;
+  flex-direction: column;
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.title {
+  margin: 1rem 0;
+}
+
+.tableCell {
+  color: var(--gray600);
+}
