+++ src/pages/account/Withdrawn/Locked/Home/Locked.tsx
@@ -2,19 +2,67 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Spinner, Typography } from '@subql/react-ui';
+import { Table, TableProps, Tag } from 'antd';
 import moment from 'moment';
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
-import { useWeb3, useWithdrawls } from '@containers';
+import { useWeb3 } from '@containers';
+import { useGetWithdrawlsQuery } from '@subql/react-hooks';
 import { defaultLockPeriod, useLockPeriod } from '@hooks';
-import { LOCK_STATUS, mapAsync, mergeAsync, notEmpty, renderAsyncArray } from '@utils';
-import { LockedList } from '../LockedList';
+import { formatEther, LOCK_STATUS, mapAsync, mergeAsync, notEmpty, renderAsyncArray } from '@utils';
+import { BigNumber } from 'ethers';
 import styles from './Locked.module.css';
+import { WithdrawalStatus } from '@subql/react-hooks/dist/graphql';
+import { TokenAmount } from '@components/TokenAmount';
+import { TableText, TableTitle } from '@subql/components';
+import { t } from 'i18next';
+import { WithdrawalFieldsFragment as Withdrawls } from '@subql/network-query';
+import { DoWithdraw } from '../DoWithdraw';
+
+interface SortedWithdrawals extends Withdrawls {
+  idx: number;
+  endAt: string;
+  lockStatus: LOCK_STATUS;
+}
+
+const dateFormat = 'MMMM Do YY, h:mm:ss a';
+
+const columns: TableProps<SortedWithdrawals>['columns'] = [
+  {
+    title: '#',
+    width: 30,
+    render: (t, r, index) => <TableText content={index + 1} />,
+  },
+  {
+    title: <TableTitle title={t('withdrawals.amount')} />,
+    dataIndex: 'amount',
+    width: 100,
+    render: (value: string) => <TokenAmount value={formatEther(value)} />,
+  },
+  {
+    title: <TableTitle title={t('withdrawals.lockedUntil')} />,
+    dataIndex: 'endAt',
+    width: 80,
+    render: (value: string) => <TableText content={moment(value).format(dateFormat)} />,
+  },
+  {
+    title: <TableTitle title={t('withdrawals.status')} />,
+    dataIndex: 'lockStatus',
+    width: 30,
+    render: (value: LOCK_STATUS) => {
+      const tagColor = value === LOCK_STATUS.UNLOCK ? 'success' : 'processing';
+      const tagContent = value === LOCK_STATUS.UNLOCK ? t('withdrawals.unlocked') : t('withdrawals.locked');
+      return <Tag color={tagColor}>{tagContent}</Tag>;
+    },
+  },
+];
 
 export const Locked: React.VFC = () => {
   const { t } = useTranslation();
   const { account } = useWeb3();
-  const withdrawals = useWithdrawls({ delegator: account || '' });
+  const withdrawals = useGetWithdrawlsQuery({
+    variables: { delegator: account || '', status: WithdrawalStatus.ONGOING, offset: 0 },
+  });
   const lockPeriod = useLockPeriod();
 
   return (
@@ -34,7 +82,28 @@ export const Locked: React.VFC = () => {
           error: (e) => <Typography>{`Error: Fail to get Indexers ${e.message}`}</Typography>,
           loading: () => <Spinner />,
           empty: () => <Typography variant="h6">{t('withdrawals.noWithdrawals')}</Typography>,
-          data: (data) => <LockedList withdrawals={data} />,
+          data: (data) => {
+            const unlockedRewards = data.filter((withdrawal) => withdrawal.lockStatus === LOCK_STATUS.UNLOCK);
+            const hasUnlockedRewards = unlockedRewards?.length > 0;
+            const withdrawalsAmountBigNumber = unlockedRewards.reduce((sum, withdrawal) => {
+              return sum.add(BigNumber.from(withdrawal.amount));
+            }, BigNumber.from('0'));
+
+            const headerTitle = `${t('withdrawals.unlockedAsset', { count: unlockedRewards?.length || 0 })}`;
+            const sortedWithdrawalsAmount = formatEther(withdrawalsAmountBigNumber);
+
+            return (
+              <div className={styles.container}>
+                <div className={styles.header}>
+                  <Typography variant="h6" className={styles.title}>
+                    {headerTitle}
+                  </Typography>
+                  <DoWithdraw unlockedAmount={sortedWithdrawalsAmount} disabled={!hasUnlockedRewards} />
+                </div>
+                <Table columns={columns} dataSource={data} rowKey="idx" />
+              </div>
+            );
+          },
         },
       )}
     </div>
