--- src/pages/account/Withdrawn/Locked/LockedList/LockedList.module.css
@@ -1,18 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-}
-
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.title {
-  margin: 1rem 0;
-}
-
-.tableCell {
-  color: var(--gray600);
-}
