--- src/pages/account/Withdrawn/Locked/LockedList/index.ts
@@ -1,4 +1,0 @@
-// Copyright 2020-2022 SubQuery Pte Ltd authors & contributors
-// SPDX-License-Identifier: Apache-2.0
-
-export * from './LockedList';
